chore: Add NPM CI recommendations
Running `npx` or `yarn dlx` in CI is non-deterministic.

`npx` will install the latest version of dependencies that match the requirements in the library's `package.json`.
This means if the library depends on version "^1.0.0" (any minor version on the 1 major) of a library, it could resolve to 1.0.0 today and 1.99.0 tomorrow.

Installing the dependency and running it via a script is preferred as we begin to treat the library like any other dependency.
That is, the resulting lockfile can be used in CI to provide a deterministic and repeatable outcome.