Copy inclusive communication doc from culture repo
I’ve found the inclusive communication guide hard to find sometimes, as
I always look for it in this repo but it is in the engineering culture
repository instead. This change copies the information into this
repository (with the intention of removing it from the other), which
will hopefully make it easier to find.